+++ src/components/trainer/parts/Portrait.vue
@@ -1,5 +1,5 @@
 <template>    
-    <div class="col-md-4">
+    <div class="col-md">
         <!-- Trainer level -->
         <div class="row">
             <div class="col-md-6">
@@ -10,19 +10,23 @@
                 <div class="text-center">Level</div>
                 <div class="text-center">{{trainer.level}}</div>
             </div>
-            <hr>
         </div>
+        <hr>
         <!-- Trainer image -->
         <div class="row">
-            <div class="text-center">Portrait</div>
-            <hr>
+            <div class="input-group my-3">
+                <span class="input-group-text">Portrait</span>
+                <select v-model="selectedSprite" class="form-select" @change="updateSprite">
+                    <option v-for="(sprite, index) in sprites" :key="index" :value="sprite.value">
+                        {{sprite.friendlyText}}
+                    </option>
+                </select>
+            </div>
             <img class="img-fluid" :src="`http://play.pokemonshowdown.com/sprites/trainers/${trainer.sprite}.png`" alt="">
-            <select v-model="selectedSprite" class="form-select my-3" @change="updateSprite">
-                <option v-for="(sprite, index) in sprites" :key="index" :value="sprite.value">
-                    {{sprite.friendlyText}}
-                </option>
-            </select>
         </div>
+        <hr>
+        <!-- Trainer Origin -->
+        <trainer-origin />
     </div>
 </template>
 
@@ -30,6 +34,7 @@
 import { getAllSprites } from '../../../api/game.api'
 import { getTrainer, setTrainer } from '../../../utils/localStorage'
 import { generateErrorModal } from '../../../utils/modalUtil'
+import TrainerOrigin from './TrainerOrigin.vue'
 export default {
     name: 'Portrait',
     data() {
@@ -40,6 +45,9 @@ export default {
             sprites: [],
             selectedSprite: ''
         }
+    },
+    components: {
+        TrainerOrigin
     },
     async beforeMount(){
         await getAllSprites()
