+++ src/components/trainer/parts/TrainerSkills.vue
@@ -1,20 +1,10 @@
 <template>
     <div class="row">
-        <div class="col-6">
+        <div class="col-12">
             <div class="text-center">Skills</div>
         </div>
-        <div class="col-6">
-            <select class="form-select text-center" name="origin" v-model="selectedOrigin" @change="updateOrigin" :disabled="isComplete">
-                <option value=""></option>
-                <option v-for="(origin, index) in origins" :key="index" :id="origin" :value="origin.replace('/', '_')">
-                    {{origin}}
-                </option>
-            </select>
-        </div>
-    </div>
-    <hr>
-    <div class="row">
-        <div class="col-12 col-6">
+        <div class="col-12">
+            <hr>
             <div class="row d-flex align-items-center">
                 <div class="col text-center">Skill Name</div>
                 <div class="col text-center">Talent 1</div>
@@ -42,26 +32,10 @@
                 </div>
             </div>
         </div>
-        <hr>
-        <div class="col-12 col-6">
-            <div class="text-center">{{selectedOrigin}} Specialty</div>
-            <div class="text-center">
-                <p>Starting Equipment: {{originData.equipment}}</p>
-                <p>Starting pokemon: {{originData.startingPokemon}}</p>
-                <div v-if="originData.feature">
-                    <p>Origin Feature: {{originData.feature.name}}</p>
-                    <p>{{originData.feature.effects}}</p>
-                    <p>
-                        Starting Savings: {{money}}¥
-                    </p>
-                </div>
-            </div>
-        </div>
     </div>
 </template>
 
 <script>
-import { getAllOrigins, getOrigin, } from '../../../api/dex.api';
 import { postLog } from '../../../api/game.api';
 import { getTrainer, setPTAActivityToken, setTrainer } from '../../../utils/localStorage'
 import { generateErrorModal } from '../../../utils/modalUtil';
@@ -70,10 +44,6 @@ export default {
     name: "TrainerSkills",
     data() {
         return {
-            origins: [],
-            selectedOrigin: '',
-            originData: {},
-            money: 0,
             trainerSkills: {},
             trainerStats: {},
             isComplete: false,
@@ -81,40 +51,14 @@ export default {
         }
     },
     beforeMount: async function(){     
-        await getAllOrigins()
-            .then(response => {
-                this.origins = response.data.results.map(item => item.name)
-            })
-            .catch(generateErrorModal);
-
         const trainer = getTrainer();
         this.trainerName = trainer.trainerName;
-        this.selectedOrigin = trainer.origin
-        this.money = trainer.money
         this.trainerSkills = trainer.trainerSkills
         this.trainerStats = trainer.trainerStats
         this.isComplete = trainer.isComplete
         await this.updateOrigin();
     },
     methods: {
-        async updateOrigin(){
-            if (this.selectedOrigin.length > 0){
-                await getOrigin(this.selectedOrigin.replace('/', '_'))
-                    .then(response => {
-                        this.updateTrainer('origin', this.selectedOrigin);
-                        this.money = response.data.savings;
-                        this.originData = response.data
-                        this.updateTrainer('money', this.money);
-                    })
-                    .catch(generateErrorModal);
-            }
-            else {
-                this.updateTrainer('origin', this.selectedOrigin);
-                this.money = 0;
-                this.originData = ''
-                this.updateTrainer('money', this.money);
-            }
-        },
         updateTalent1(index){
             this.trainerSkills[index].talent1 = !this.trainerSkills[index].talent1
             this.updateTrainer('trainerSkills', this.trainerSkills);
@@ -139,7 +83,7 @@ export default {
             .then(response => {
                 setPTAActivityToken(response.headers['pta-activity-token']);
             })
-
+            .catch(generateErrorModal)
         }
     }
 }
